+++ src/service/tantivy/reader/mod.rs
@@ -17,19 +17,14 @@ mod parquet;
 #[cfg(all(feature = "vortex", feature = "enterprise"))]
 mod vortex;
 
-use std::io::Cursor;
 #[cfg(all(feature = "vortex", feature = "enterprise"))]
 use std::ops::Range;
 
-use ::parquet::arrow::{
-    ParquetRecordBatchStreamBuilder,
-    arrow_reader::{ParquetRecordBatchReader, ParquetRecordBatchReaderBuilder},
-};
 use arrow::{error::ArrowError, record_batch::RecordBatch};
 use bytes::Bytes;
-use config::{FileFormat, get_batch_size, utils::parquet::RecordBatchStream};
-use futures::TryStreamExt;
-use parquet::parquet_projection;
+use config::{FileFormat, utils::parquet::RecordBatchStream};
+
+pub(super) type RecordBatchIter = Box<dyn Iterator<Item = Result<RecordBatch, ArrowError>>>;
 
 pub(super) enum ChunkSelector {
     Parquet(usize), // row_group id
@@ -45,18 +40,7 @@ pub(super) async fn file_stream(
     projection: Option<&[String]>,
 ) -> Result<RecordBatchStream, anyhow::Error> {
     match file_format {
-        FileFormat::Parquet => {
-            let builder = ParquetRecordBatchStreamBuilder::new(Cursor::new(data)).await?;
-            let mut builder = builder.with_batch_size(get_batch_size());
-            if let Some(mask) =
-                parquet_projection(builder.schema(), builder.metadata(), projection)?
-            {
-                builder = builder.with_projection(mask);
-            }
-            let reader = builder.build()?;
-            let stream: RecordBatchStream = Box::pin(reader.map_err(ArrowError::from));
-            Ok(stream)
-        }
+        FileFormat::Parquet => parquet::scan_parquet_async(data, projection).await,
         #[cfg(all(feature = "vortex", feature = "enterprise"))]
         FileFormat::Vortex => vortex::scan_vortex_async(data, projection).await,
         #[cfg(not(all(feature = "vortex", feature = "enterprise")))]
@@ -71,42 +55,14 @@ pub(super) fn chunk_iter(
     selector: ChunkSelector,
     data: Bytes,
     projection: Option<&[String]>,
-) -> Result<ChunkBatchIter, anyhow::Error> {
+) -> Result<RecordBatchIter, anyhow::Error> {
     match selector {
         ChunkSelector::Parquet(row_group_id) => {
-            let builder = ParquetRecordBatchReaderBuilder::try_new(data)?;
-            let mut builder = builder.with_batch_size(get_batch_size());
-            if let Some(mask) =
-                parquet_projection(builder.schema(), builder.metadata(), projection)?
-            {
-                builder = builder.with_projection(mask);
-            }
-            let reader = builder.with_row_groups(vec![row_group_id]).build()?;
-            Ok(ChunkBatchIter::Parquet(reader))
+            parquet::scan_parquet_row_group(data, projection, row_group_id)
         }
         #[cfg(all(feature = "vortex", feature = "enterprise"))]
         ChunkSelector::Vortex(row_range) => {
-            let iter = vortex::scan_vortex_row_range(data, projection, row_range)?;
-            Ok(ChunkBatchIter::Vortex(iter))
-        }
-    }
-}
-
-/// Unified sync iterator over parquet or vortex record batches, returned by [`chunk_iter`].
-pub(super) enum ChunkBatchIter {
-    Parquet(ParquetRecordBatchReader),
-    #[cfg(all(feature = "vortex", feature = "enterprise"))]
-    Vortex(vortex::VortexRowRangeIter),
-}
-
-impl Iterator for ChunkBatchIter {
-    type Item = Result<RecordBatch, ArrowError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            Self::Parquet(r) => r.next(),
-            #[cfg(all(feature = "vortex", feature = "enterprise"))]
-            Self::Vortex(v) => v.next(),
+            vortex::scan_vortex_row_range(data, projection, row_range)
         }
     }
 }
